refactor(llm-train): merge over-fragmented modules (9 -> 7)
Zero behavior change; consolidates thin single-consumer modules that the
planning-time one-file-per-task split over-fragmented.

- Fold chunked_logprob.py (kernels) into chunked_policy.py -- the kernels had
  exactly one consumer.
- Fold reward.py (TradingReward) into trading_env.py -- used only by
  TradingRewardParser in the same file.
- Collapse the one-entry loss "registry" in losses.py to a direct if/elif; keep
  the callable-factory extension seam.

Full suite green (1808 passed, 15 skipped).
