+++ tests/llm/train/test_reward.py
@@ -1,5 +1,5 @@
 import pytest
-from torchtrade.llm.train.reward import TradingReward
+from torchtrade.llm.train.trading_env import TradingReward
 
 
 class _FakeEnv:
